Added basic --help option to display attributes
If you receive this useful utility in binary form, there is no obvious way to know the command line options. A simplistic help menu was added on --help so that it's obvious how to indicate the location of libjlinkarm.dylib if you're on OS X, etc., since the current implementation on searches for .so.